Refactor saving preview data completely to store
[MAILPOET-2743]
